issue #101: migrated RandomGenerator to outbound port and fixed id generation ownership
