+++ src/main/kotlin/org/agrfesta/sh/api/core/application/ports/outbounds/RandomGenerator.kt
@@ -1,0 +1,10 @@
+package org.agrfesta.sh.api.core.application.ports.outbounds
+
+import java.util.UUID
+
+interface RandomGenerator {
+
+    fun string(): String
+    fun uuid(): UUID
+
+}
