+++ src/main/kotlin/org/agrfesta/sh/api/core/application/usecases/CreateAreaService.kt
@@ -3,9 +3,9 @@ package org.agrfesta.sh.api.core.application.usecases
 import arrow.core.Either
 import org.agrfesta.sh.api.core.application.ports.inbounds.CreateAreaUseCase
 import org.agrfesta.sh.api.core.application.ports.outbounds.areas.AreasRepository
+import org.agrfesta.sh.api.core.application.ports.outbounds.RandomGenerator
 import org.agrfesta.sh.api.core.domain.areas.AreaDto
 import org.agrfesta.sh.api.core.domain.failures.AreaCreationFailure
-import org.agrfesta.sh.api.utils.RandomGenerator
 import org.springframework.stereotype.Service
 
 @Service
