+++ src/main/kotlin/org/agrfesta/sh/api/core/application/usecases/RefreshDevicesService.kt
@@ -7,7 +7,7 @@ import arrow.core.right
 import org.agrfesta.sh.api.core.application.ports.inbounds.RefreshDevicesUseCase
 import org.agrfesta.sh.api.core.application.ports.outbounds.devices.DevicesProvider
 import org.agrfesta.sh.api.core.application.ports.outbounds.devices.DevicesRepository
-import org.agrfesta.sh.api.utils.RandomGenerator
+import org.agrfesta.sh.api.core.application.ports.outbounds.RandomGenerator
 import org.agrfesta.sh.api.core.domain.devices.Device
 import org.agrfesta.sh.api.core.domain.devices.DeviceStatus
 import org.agrfesta.sh.api.core.domain.devices.RefreshDevicesResult
