+++ src/main/kotlin/org/agrfesta/sh/api/infrastructure/utils/RandomGeneratorImpl.kt
@@ -1,0 +1,11 @@
+package org.agrfesta.sh.api.infrastructure.utils
+
+import org.agrfesta.sh.api.core.application.ports.outbounds.RandomGenerator
+import org.springframework.stereotype.Service
+import java.util.UUID
+
+@Service
+class RandomGeneratorImpl : RandomGenerator {
+    override fun string(): String = uuid().toString()
+    override fun uuid(): UUID = UUID.randomUUID()
+}
